[core] Extract WorkTaskImpl from run_loop.hpp
